net, netbinding: Extract net-binding logic to the network package
Network specific logic on how to create network binding sidecars is
moved from the services package to the network package.

With this change, adding new network binding sidecars should not require
any change in the virt-controller sub-packages.
