+++ pkg/network/netbinding/netbinding_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package services_test
+package netbinding_test
 
 import (
 	"fmt"
@@ -31,7 +31,8 @@ import (
 	v1 "kubevirt.io/api/core/v1"
 
 	"kubevirt.io/kubevirt/pkg/hooks"
-	"kubevirt.io/kubevirt/pkg/virt-controller/services"
+	"kubevirt.io/kubevirt/pkg/network/netbinding"
+
 	"kubevirt.io/kubevirt/tests/libvmi"
 )
 
@@ -44,8 +45,6 @@ var _ = Describe("Network Binding", func() {
 		testBindingName2  = "binding2"
 		testSidecarImage2 = "image2"
 		testNetworkName3  = "net1"
-
-		shouldSucceed = true
 	)
 
 	Context("binding plugin sidecar list", func() {
@@ -55,7 +54,7 @@ var _ = Describe("Network Binding", func() {
 					Binding: bindings,
 				},
 			}
-			sidecars, err := services.NetBindingPluginSidecarList(vmi, config, nil)
+			sidecars, err := netbinding.NetBindingPluginSidecarList(vmi, config, nil)
 			Expect(err).ToNot(HaveOccurred())
 			Expect(sidecars).To(ConsistOf(expectedSidecars))
 		},
@@ -104,7 +103,7 @@ var _ = Describe("Network Binding", func() {
 			config := &v1.KubeVirtConfiguration{
 				NetworkConfiguration: &v1.NetworkConfiguration{},
 			}
-			_, err := services.NetBindingPluginSidecarList(vmi, config, record.NewFakeRecorder(1))
+			_, err := netbinding.NetBindingPluginSidecarList(vmi, config, record.NewFakeRecorder(1))
 			Expect(err).To(HaveOccurred())
 		})
 	})
@@ -122,7 +121,7 @@ var _ = Describe("Network Binding", func() {
 				},
 			}
 
-			Expect(services.SlirpNetBindingPluginSidecar(vmi, config, fakeRecorder)).To(Equal(hooks.HookSidecar{
+			Expect(netbinding.SlirpNetBindingPluginSidecar(vmi, config, fakeRecorder)).To(Equal(hooks.HookSidecar{
 				ImagePullPolicy: k8scorev1.PullIfNotPresent,
 				Image:           "kubevirt/network-slirp-plugin",
 			}))
@@ -133,8 +132,8 @@ var _ = Describe("Network Binding", func() {
 				fakeRecorder := record.NewFakeRecorder(1)
 				vmi := v1.NewVMI("test", "1234")
 
-				Expect(services.SlirpNetBindingPluginSidecar(vmi, config, fakeRecorder)).To(Equal(hooks.HookSidecar{
-					Image: services.DefaultSlirpPluginImage,
+				Expect(netbinding.SlirpNetBindingPluginSidecar(vmi, config, fakeRecorder)).To(Equal(hooks.HookSidecar{
+					Image: netbinding.DefaultSlirpPluginImage,
 				}))
 			},
 			Entry("has no network configuration set", &v1.KubeVirtConfiguration{}),
@@ -165,15 +164,15 @@ var _ = Describe("Network Binding", func() {
 			fakeRecorder := record.NewFakeRecorder(1)
 			vmi := v1.NewVMI("test", "1234")
 
-			Expect(services.SlirpNetBindingPluginSidecar(vmi, &v1.KubeVirtConfiguration{}, fakeRecorder)).To(Equal(hooks.HookSidecar{
-				Image: services.DefaultSlirpPluginImage,
+			Expect(netbinding.SlirpNetBindingPluginSidecar(vmi, &v1.KubeVirtConfiguration{}, fakeRecorder)).To(Equal(hooks.HookSidecar{
+				Image: netbinding.DefaultSlirpPluginImage,
 			}))
 
 			Expect(fakeRecorder.Events).To(HaveLen(1))
 			event := <-fakeRecorder.Events
 			Expect(event).To(Equal(
 				fmt.Sprintf("Warning %s no Slirp network binding plugin image is set in Kubevirt config, using '%s' sidecar image for Slirp network binding configuration",
-					services.UnregisteredNetworkBindingPluginReason, services.DefaultSlirpPluginImage),
+					netbinding.UnregisteredNetworkBindingPluginReason, netbinding.DefaultSlirpPluginImage),
 			))
 		})
 	})
