+++ pkg/virt-controller/services/template_test.go
@@ -136,10 +136,6 @@ var _ = Describe("Template", func() {
 					func(vmi *v1.VirtualMachineInstance, _ *v1.KubeVirtConfiguration) (hooks.HookSidecarList, error) {
 						return hooks.UnmarshalHookSidecarList(vmi)
 					}),
-				WithSidecarCreator(
-					func(vmi *v1.VirtualMachineInstance, kvc *v1.KubeVirtConfiguration) (hooks.HookSidecarList, error) {
-						return NetBindingPluginSidecarList(vmi, kvc, record.NewFakeRecorder(1))
-					}),
 			)
 			// Set up mock clients
 			networkClient := fakenetworkclient.NewSimpleClientset()
@@ -2936,48 +2932,6 @@ var _ = Describe("Template", func() {
 			Expect(pod.Spec.Containers).To(HaveLen(2))
 			Expect(pod.Spec.Containers[1].Image).To(Equal(testHookSidecar.Image))
 			Expect(pod.Spec.Containers[1].ImagePullPolicy).To(Equal(testHookSidecar.ImagePullPolicy))
-		})
-
-		Context("with slirp interface", func() {
-			It("should add slirp network binding hook sidecar container with default image when no image is registered", func() {
-				vmi := v1.VirtualMachineInstance{ObjectMeta: metav1.ObjectMeta{
-					Name: "testvmi", Namespace: "default", UID: "1234",
-				}}
-				vmi.Spec.Networks = append(vmi.Spec.Networks, *v1.DefaultPodNetwork())
-				vmi.Spec.Domain.Devices.Interfaces = append(vmi.Spec.Domain.Devices.Interfaces, *v1.DefaultSlirpNetworkInterface())
-
-				config, kvInformer, svc = configFactory(defaultArch)
-				pod, err := svc.RenderLaunchManifest(&vmi)
-				Expect(err).ToNot(HaveOccurred())
-
-				Expect(pod.Spec.Containers).To(HaveLen(2))
-				Expect(pod.Spec.Containers[1].Image).To(Equal("quay.io/kubevirt/network-slirp-binding:20230830_638c60fc8"))
-				Expect(pod.Spec.Containers[1].ImagePullPolicy).To(Equal(kubev1.PullPolicy("IfNotPresent")))
-			})
-			It("should add slirp network binding hook sidecar container using registered image", func() {
-				const slirpPluginSidecarImage = "kubevirt/network-slirp-binding:v1"
-				testBindingPluginNetConf := &v1.NetworkConfiguration{Binding: map[string]v1.InterfaceBindingPlugin{"slirp": {SidecarImage: slirpPluginSidecarImage}}}
-
-				config, kvInformer, svc = configFactory(defaultArch)
-				kvConfig := kv.DeepCopy()
-				kvConfig.Spec.Configuration.NetworkConfiguration = testBindingPluginNetConf
-				testutils.UpdateFakeKubeVirtClusterConfig(kvInformer, kvConfig)
-
-				vmi := v1.VirtualMachineInstance{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "testvmi", Namespace: "default", UID: "1234",
-					},
-				}
-				vmi.Spec.Networks = append(vmi.Spec.Networks, *v1.DefaultPodNetwork())
-				vmi.Spec.Domain.Devices.Interfaces = append(vmi.Spec.Domain.Devices.Interfaces, *v1.DefaultSlirpNetworkInterface())
-
-				pod, err := svc.RenderLaunchManifest(&vmi)
-				Expect(err).ToNot(HaveOccurred())
-
-				Expect(pod.Spec.Containers).To(HaveLen(2))
-				Expect(pod.Spec.Containers[1].Image).To(Equal(slirpPluginSidecarImage))
-				Expect(pod.Spec.Containers[1].ImagePullPolicy).To(Equal(kubev1.PullPolicy("IfNotPresent")))
-			})
 		})
 
 		Context("with pod networking", func() {
