+++ pkg/virt-controller/watch/BUILD.bazel
@@ -32,6 +32,7 @@ go_library(
         "//pkg/monitoring/vmistats:go_default_library",
         "//pkg/monitoring/vmstats:go_default_library",
         "//pkg/network/namescheme:go_default_library",
+        "//pkg/network/netbinding:go_default_library",
         "//pkg/network/sriov:go_default_library",
         "//pkg/network/vmispec:go_default_library",
         "//pkg/service:go_default_library",
