+++ tests/network/BUILD.bazel
@@ -33,6 +33,7 @@ go_library(
         "//pkg/apimachinery/patch:go_default_library",
         "//pkg/cloud-init:go_default_library",
         "//pkg/network/istio:go_default_library",
+        "//pkg/network/netbinding:go_default_library",
         "//pkg/network/setup:go_default_library",
         "//pkg/network/vmispec:go_default_library",
         "//pkg/util:go_default_library",
